#15 collectionViewDelegate 수정
