--- uckgisagi-iOS/uckgisagi-iOS/Screen/Home/UserProfileTableViewHeader.swift
@@ -1,70 +1,0 @@
-//
-//  UserProfileTableViewHeader.swift
-//  uckgisagi-iOS
-//
-//  Created by Sojin Lee on 2022/11/01.
-//
-
-import UIKit
-
-import SnapKit
-
-class UserProfileTableViewHeader: UITableViewHeaderFooterView {
-    var collectionView: UICollectionView = {
-        var layout = UICollectionViewFlowLayout()
-        layout.minimumLineSpacing = 0
-        layout.scrollDirection = .horizontal
-        layout.sectionInset = .zero
-        
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.contentInset = UIEdgeInsets(top: .zero, left: 20, bottom: .zero, right: .zero)
-        return collectionView
-    }()
-    
-    override init(reuseIdentifier: String?) {
-        super.init(reuseIdentifier: reuseIdentifier)
-        setLayouts()
-        setupCollectionView()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func setLayouts() {
-        addSubview(collectionView)
-        
-        collectionView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-    }
-}
-
-extension UserProfileTableViewHeader: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
-    func setupCollectionView() {
-        collectionView.delegate = self
-        collectionView.dataSource = self
-        collectionView.register(UserProfileCollectionViewCell.self, forCellWithReuseIdentifier: UserProfileCollectionViewCell.identifier)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 5 // MARK: - 서버에서 받는 값으로 수정하기
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: UserProfileCollectionViewCell.identifier, for: indexPath) as? UserProfileCollectionViewCell else {
-            return UICollectionViewCell()
-        }
-        if indexPath.row == 4 {
-            cell.configureLastCell()
-        } else {
-            cell.configureProfile(name: "뭉", isFriend: false) // MARK: - 서버에서 받는 값으로 수정하기
-        }
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: 48, height: 54)
-    }
-}
